+++ app/services/forest_liana/operator_value_parser.rb
@@ -59,18 +59,28 @@ module ForestLiana
     end
 
     def self.get_field_name(field, resource)
-      if field.split(':').size < 2
-        "#{resource.quoted_table_name}." +
-        "#{ActiveRecord::Base.connection.quote_column_name(field)}"
-      else
+      if self.is_belongs_to(field)
         association = field.split(':')[0].pluralize
         "#{ActiveRecord::Base.connection.quote_column_name(association)}." +
         "#{ActiveRecord::Base.connection.quote_column_name(field.split(':')[1])}"
+      else
+        "#{resource.quoted_table_name}." +
+        "#{ActiveRecord::Base.connection.quote_column_name(field)}"
       end
     end
 
     def self.format_value(resource, field, value)
-      column = resource.columns.find { |c| c.name == field }
+      if self.is_belongs_to(field)
+        fields = field.split(':')
+        columns = resource.reflect_on_association(fields[0]).klass.columns
+        field_name = fields[1]
+      else
+        columns = resource.columns
+        field_name = field
+      end
+
+      column = columns.find { |column| column.name == field_name }
+
       if column.type == :boolean
         ForestLiana::AdapterHelper.cast_boolean(value)
       else
@@ -78,5 +88,8 @@ module ForestLiana
       end
     end
 
+    def self.is_belongs_to(field)
+      field.split(':').size >= 2
+    end
   end
 end